requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==8.0.1
discord.py==1.7.3
docopt==0.6.2
elasticsearch==7.13.1
fandom-py==0.2.1
geographiclib==1.52
geopy==2.1.0
googlesearch-python==1.0.1
googletrans==3.1.0a0
googletransx==2.4.2
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
memeapi==0.2.3
multidict==5.1.0
numpy==1.21.0
pandas==1.2.5
pip-autoremove==0.9.1
pycares==4.0.0
pycparser==2.20
pyee==8.1.0
pyppeteer==0.2.5
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
tabulate==0.8.9
tqdm==4.61.2
typing-extensions==3.10.0.0
udpy==2.0.0
urllib3==1.26.5
websockets==9.1
wikipedia==1.4.0
yarl==1.6.3
youtube-search==2.1.0
Pillow==8.3.1
